Use cmath instead of math.h (#597)
* Use cmath instead of math.h

- use "std" namespace math functions (instead of f* versions) to make sure we get the correct overload
- add "std::" namespace to some math functions to ensure we get the correct overload (only did the obvious ones in the files I was changing the includes - needs to be done throughout the code base)
- use proper types for some initializations
- remove math.h completely on cases where it is unnecessary
- remove _USE_MATH_DEFINES check in qHoughNormalsDialog.h (if this is still necessary for MSVC, I'd suggest adding -D_USE_MATH_DEFINES globally in CMAkeSetCompilerOptions.cmake)

* Replace qMax with std::max

* Fixes based on review

+++ libs/qCC_db/ccColorTypes.h
@@ -25,7 +25,7 @@
 #include <QColor>
 
 //system
-#include <math.h> //for modf
+#include <cmath>
 #include <random>
 
 //! Default color components type (R,G and B)
@@ -216,12 +216,12 @@ namespace ccColor
 		**/
 		static Rgb hsv2rgb(float H, float S, float V)
 		{
-			double hi = 0;
-			double f = modf(H / 60.0, &hi);
+			float hi = 0;
+			float f = std::modf(H / 60.0f, &hi);
 
-			float l = static_cast<float>(V*(1.0 - S));
-			float m = static_cast<float>(V*(1.0 - f*S));
-			float n = static_cast<float>(V*(1.0 - (1.0 - f)*S));
+			float l = V*(1.0f - S);
+			float m = V*(1.0f - f*S);
+			float n = V*(1.0f - (1.0f - f)*S);
 
 			Rgbf rgb(0, 0, 0);
